feat: add breakglass capabilities (#154)
This adds the ability to create AoD requests for breakglass scenarios.
By adding the `aod-breakglass` label to an AoD pr, it will automatically
be applied.

This approach was selected over requiring an issue link as this creates
less toil. The justification can be provided in the PR body rather than
in a separate issue which just adds an extra step on a separate screen.